Move jsonStringify test from atomic/string.spec.ts
- Created packages/run-types/src/jitCompilers/json/jsonStringify.spec.ts
- Moved test from packages/run-types/src/runType/atomic/string.spec.ts:34-40
- All tests passing
