+++ packages/run-types/src/jitCompilers/json/jsonStringify.spec.ts
@@ -1,0 +1,26 @@
+/* ########
+ * 2024 mion
+ * Author: Ma-jerez
+ * License: MIT
+ * The software is provided "as is", without warranty of any kind.
+ * ######## */
+
+import {runType} from '../../lib/runType';
+import {JitFunctions} from '../../constants.functions';
+
+describe('jsonStringify compilation tests', () => {
+    // Tests moved from various spec files under packages/run-types/src/runType/
+
+    // Moved from packages/run-types/src/runType/atomic/string.spec.ts:34-40
+    {
+        const rt = runType<string>();
+
+        it('json stringify', () => {
+            const jsonStringify = rt.createJitFunction(JitFunctions.jsonStringify);
+            const fromJsonVal = rt.createJitFunction(JitFunctions.fromJsonVal);
+            const typeValue = 'hello';
+            const roundTrip = fromJsonVal(JSON.parse(jsonStringify(typeValue)));
+            expect(roundTrip).toEqual(typeValue);
+        });
+    }
+});
